break: move `utils.Encoder` -> `buffer.Encoder`
